requirements.txt
python-dotenv
pillow
google-genai
